requirements.txt
pandas==1.4.4
numpy==1.21.5
requests==2.28.1
psycopg2-binary==2.9.6
mysql-connector-python==9.4.0